Replace prompt mark shell escape string
Part of #651, this removes a shell prompt escape string, making it
easier to work with terminal titles, as well as making it possible to
customize the prompt mark for the root user. This patch does not do
that, as the functionality has remained the same, but it would now be
possible to add a LP_MARK_ROOT.

This patch should have no functional difference, as the checks that Bash
and Zsh did in the prompt escape are exactly the same that Liquidprompt
is doing: comparing EUID to 0.