Replace hostname shell prompt escape with env vars
Again further work on #651. This time it is simple: both Bash and Zsh
set the exact same hostname that they use for their prompt escape
sequences into HOSTNAME and HOST, respectively.

Then it is a simple matter to replicate their "short" version by
truncating the string to the first period.