Remove backslash escaping from Zsh __lp_escape()
This was added in commit 74e14fe0a7acb6c912d8fc800eca0b39d2cdf069. I
don't understand why, except that maybe a single backslash display would
look like an escape code for a control character, like \n. But that is
not the case for Bash, so why the difference?

Zsh only uses percents (%) to escape things in the prompt, where Bash
only uses bashslashes. This changes makes the same input string print
the same way in the prompt for both shells.