Fix LP_COLORMAP reverting to default value
When the LP_COLORMAP config option contained a first element that was
the empty string, lp_activate() would think it was empty, and set
LP_COLORMAP back to the default value.

This was identified by @milouse in a comment here:
https://github.com/nojhan/liquidprompt/pull/628#discussion_r611182040

The issue is not severe, as any value set in the config will always be
taken (as it is set last). But a value set in the environment, either
manually or in a shell RC file would be overridden (as long as the first
element was the empty string).

Commit e121179d1bb943ec3673e451deac2a0577adbb75 has a good explaination
of why the ${array[@]+"${array[@]}"} form is used. This unfortunately
must be used _anywhere_ an array could be empty or have any empty
strings as values.