Fix bug in powerlines for Zsh and Bash-3.2
Undeclared local variables are the worst. When declaring a local
variable, Zsh treats it as set (to the empty) string, while Bash (more
correctly, in my opinion) treats it as unset. This meant that this check
of if the first powerline section had happened yet would always return
true on Zsh, showing an ugly arrow at the start of the prompt. On
Bash-3.2 (with set -u) it would just crash, which is how I noticed the
bug.

The solution is to never rely on a local being set or not, but to look
at the actual value. In this case check for the empty string (with a set
-u safety).