Fix temperature erroring if tool returns no data
Rarely, sensors or acpi can return no temperature data. In this case,
the data function would return false and 0, but the detect function
would error if set -u was enabled.

Harden all of the four temperature functions against this, and greatly
increase the test coverage to cover these four functions, as well as
adding a specific case for the tools returning no data.

This also allows for the temperature functions to return values less
than 0, if somehow a device has all temp sensors returning less than
that.

Mention in the function documentation that it is possible for the
function to return a not set string.

Fixes 670