Fix bash __lp_escape() for backtick expansion
For some reason, Bash ignores a '\$' sequence when the promptvars option
is off, printing it as '$'. It does not do the same for a backtick
escaped: '\`' is printed as '\`'. To prevent this issue, detect if the
option is on, and if so, escape both '$' and '`'.

Running 'shopt' is actually faster than either a [[ ]] or (( ))
expression.