Add Kubernetes context default namespace
Add data gathering for the configured default namespace of the current
kubectl context. Display that data as a string directly after the
context string, separated by a colon.

While this is a separate chunk of data from the context, it is actually
part of the context definition, and so it makes sense to display it
along with the context.

While most Kubernetes objects are subject to the strict DNS name
restrictions, kubectl contexts never have their name server-side, and so
do not have the same restrictions. This means a context name can include
any characters, meaning we have no safe way to parse it out of a data
string. I wanted to use:

line=$(kubectl config view --minify --output 'jsonpath={.current-context}{" "}{..namespace}')
IFS=' ' read kubernetes_context kubernetes_namespace <<<"$line"

But since spaces can be in a context name, this would break.