Remove ubuntu-16.04 env from tests
Github has removed this version of Ubuntu VM, as it is no longer
supported by the upstream.

Change the targets to the rolling versions of "latest", to prevent
needing to make these changes again. My original logic was to run on as
many OSs as possible, but in reality all of these versions act so
similar in terms of shell functionality that it is kinda pointless. One
VM of each OS should be fine.