Fix zsh glob error
Zsh by default will error if a glob matches nothing (a smart default).
We need it to not error, and even better, to not use the Bash default of
using the pattern. Set nullglob to expand it to nothing.

Don't set the Bash version of nullglob in liquidprompt, since Bash
doesn't do local options well, and not at all until version 4.4. It is
safe to expand to a single string with the "*" in it, since the first
check we do is for the file existing, which it will not.

Add the sysfs power supply interface to the external-tool-tester.sh.