Fix picking acpi battery backend when invalid
With the removal of the command -v check, anything returned on stdout
would make acpi be a valid backend. This can happen in the
command_not_found_handle() function is defined.

Also don't pick a backend if the data function returns an error code.

Further strictness could be to check that the string returned is a valid
int, but that is probably overkill.

Fixes #690