Fix slowdown from undefined VCS functions
Without these functions defined, each prompt invocation would force the
shell to check if the command was valid, and generate an error message
if it was not. We know that it won't be, so we can shortcut that by
setting it as such.

More importantly, if the Bash function command_not_found_handle() is
set, it will be executed for each function that is not set. For Git or
HG, this is only one. But for SVN, it is enough to lock the prompt for a
while.

Fixes #696