Add config option for changing hg command
Mercurial is written in Python 2, so it was always slow. There is work
ongoing to rewrite it in Rust, but as the replacement is not at feature
pairity yet, they are shipping the binary as 'rhg'. Since some might
want to use different binaries for different things, give a config
option to pick which binary to use for prompt data gathering.

Resolves #701