Add and fix hostname display methods
Since it was added, the LP_ENABLE_FQDN option did not do what its name
claimed to do. It instead provided the full hostname. This was not
exactly a bug, as it was using the options provided by Bash and Zsh, but
it did not provide the "fully qualified domain name", and so at best it
was misleading.

The solution is to add a new config option that supports short, full,
and fqdn options to let the user select which they want. The new fqdn
option is provided by `hostname -f`, as long as the user's OS supports
that.

This deprectates the LP_ENABLE_FQDN option.

Related, there was a request for a "pretty" hostname option as well, so
that has also been added. Only MacOS and Systemd support such a concept
as far as I can tell.

Resolves #472
Resolves #596