Fix some terminal emulators showing as telnet
Currently the Guake terminal emulator is being detected as a telnet
session. The current detection method for telnet is to use the `who`
binary to check for a connection string. Connecting with telnet will
show something in the comment field like "(localhost)".

The issue is that _any_ connection into a terminal will show something
in that comment, like "(:0)" for a desktop environment, or "(m)" for
Quake, or "(tmux)" for Tmux. There really is no way to detect which is a
hostname as opposed to some local description.

Since this has often been an issue, this change is to remove the
dependency on `who`, and look instead at $REMOTEHOST. This is a not very
well documented environment variable that is set when connecting with
telnet. It seems that it is not a standard at all, but some telnet
daemons and shells support setting it. Since we have seen a lot of
reports of false positives detecting telnet, I would rather be on the
side of false negative and miss a few telnet connections. But telnet is
so rare now, so it probably wouldn't matter anyway.

Resolves #483
Closes #556
Closes #562