Fix unset bug under zsh
This was unlikely to be hit, but it was possible, and was hit in the
tests.

Weirdly, it fails because $var expands to the first element of the array
if $var is an array, and since if it is and it is an empty array,
${#var} expands to length of the first element of the array, which since
there is not one, it fails because it is unset.

The correct syntax is ${#var[@]}, which is length of the array.