Fix Git worktree detection and add GIT_DIR support
The work in 2.0 for VCS detection broke a lot of stuff. I've fixed most
of it to this point (thanks to everyone who reported issues), but there
are a few weird quirks here or there.

The odd one is Git. It supports an environment variable $GIT_DIR that
makes Git run the repo from that folder instead of a .git/ directory
somewhere up the tree. This is rarely used, but VCSH does use it.

In fixing this, I noticed that our state detection while using Git
worktrees was also broken, but that is a simple fix.

This also pulls out the detection of VCSH and Git-SVN out of
_lp_smart_mark(), and into _lp_find_vcs().