Make VCSH print $VCSH_DIRECTORY
Entering a VCSH environment without showing that you entered is a bit
confusing, only indicator is that the branch name for git shows up.
Liquidprompt already shows virtualenvs, so vcsh should be treated
similarly.

This may not be the right place to display this, but it is by far the
simpliest.

Resolves #363