Add displaying selected Node.js version using NVM
NVM does not expose the version in any environment variables, which
makes this annoying.

This also doesn't quite fall under development environments, since it
isn't an environment thing, and it isn't directory specific either.

But this PR was already opened, and it was simple to add to the already
existing nodeenv functionality.

Co-authored-by: Mark Korondi <korondi.mark@gmail.com>

Closes #561
Related #298