Fix external-tool-tester.sh on OS without release
On filesystems without anything matching /etc/*release or /etc/issue*,
the nullglob option that was set would make the shell expand the command
to `cat `, which tells cat to wait for input on stdin. This causes the
script to hang.

The fix is to pass /dev/null to stdin so that cat knows nothing is
coming, so it can exit right away. This does make the printf statement
not run, but whatever.