Add support for Bash 5.1+ PROMPT_COMMAND array
I held off on this one for a long time, since there is no good way to
test if a variable is an array or not. But it turns out that does not
really matter. You can treat the variable like it is an array, and it
will just become one.

This should not have any drawbacks that I can see. It also doesn't
matter if __lp_set_prompt is the first element in the array, as all
commands in the array get the correct `$?` value.

I still think that Bash should have made a new variable, like
PROMPT_COMMANDS, instead of this backwards breaking change.

Also rework a bit of how to delete an element from the array. This
should be faster too.

Resolves #684