Improve method for stripping shell escape sequences
This has always been used for printing things to the terminal title, but
in the future we have more plans that require this (counting line
length). The issue is that by forking out to sed, this is slow. It is
actually two forks; one for the call to _lp_as_text (as it is in a
subshell), and a second to sed.

__lp_strip_escapes() removes both forks. It strips everything between
the shell formatting escape sequences, and also removes extra
backslashes that we added previously with __lp_escape(). This gives us
pure text, though it does not remove control characters.

Since it can undo all of the shell escaping, we have no more need for
lp_hostname_raw or lp_username_raw. They were added to make the title
generation much easier, but now we can strip it after the fact. We will
keep lp_path though, as it is formatted differently since it does not
have the custom path separators.

In the best case senario, these fork removals have cut prompt generation
time in half!