Add test for activate function
The ruby test ended up somewhat testing this, but we really need a test
just for this, and for the ruby test to not depend on it.

Note this isn't a very good test, pretty much only checking if there are
any unset variables. Using set -e just does not work, as there are
statements that return non-zero by design.