Fix error messages on sysfs with buggy drivers
Some devices might setup a device file, but fail to bind a backing
device to it. This makes the file pass all tests we can throw at it, but
attempting to open it returns a "No such device or address" error. This
seems to be a driver bug.

There is no way to test for this without opening it, so try to open it,
and if it fails, ignore the device.

The fix to the tests is because the read will exit as soon as it hits
EOF, even if it is part way through reading a line. But the real files
will always include a terminating newline, so this is a safe test.

Resolves #713