Fix new dev env section not clearing values
The new dev environments section pulls from the string outputs of all
the dev environment functions. But because it only looks at the strings,
not the return values (since those are captured elsewhere), it missed
when the functions started returning false. Since the functions don't
clear their string return value in this case, it kept on printing the
value.

The fix is for those functions to do exactly that; clear their return
strings. It is not the cleanest, nor the most consistent (since most
data functions do not do this), but it works for now.

This does not break any API or contract, since the documentation states
that when a data function returns false, the return value is undefined.