Fix error meaning feature enabled by error feature
Maybe had been originally intended to work this way, but it means the
default as specified in the documentation is incorrect. Since they do
not output in the same function or even section, it makes sense not to
tie them together.
Must enable the feature in the tests that assumed it would be on.