Fix themes relying on _lp_fill() for theme-preview
Themes that used __lp_string_escapes as a part of printing the prompt,
usually using the _lp_fill() function, would not work correctly in the
tools/theme-preview.sh script. This was due to an earlier assumption
that the escape sequences were only for shell interpretation in PS1.
This is true, but we also use them internally to know where to removing
formatting. The fix is simple, remove those strings as the very last
step before printing.

Fixes #753