Fix Poetry virtualenvs showing "prompt ="
The file format for pyvenv.cfg is not very well defined, and it seems
that the line could be any of the following:

* prompt = venv
* prompt=venv
* prompt='venv'
* prompt = 'venv'

This gives us a lot of cases to cover, as well as making parsing the
file very hard. The new _grep_fields() function can't be used, as it
expects a single defined delimiter, which we don't have.

Fixes #760