Fix runtime formatting on Zsh if SECONDS is float
On Zsh it is possible to set the type of SECONDS to a float. If this is
done before loading liquidprompt, _LP_RUNTIME_SECONDS will take the type
of SECONDS, and will then display weirdly with partial seconds.

The fix is to force _LP_RUNTIME_SECONDS to be an int.

Fixes #790