Fix Git lookups when in Git worktree
Git SVN detection and the new remote branch detection were broken when
in a worktree. Worktrees have two Git dirs: one is the main dir, which
holds the DB and config. The other is the worktree specific dir, which
holds directory specific information, like index and status like merge
or rebase in progress.