fix(_lp_fill): avoid costly strip_escape on filling sequence
The previous behavior was to allow escaped sequences in the *fillchars* argument of _lp_fill.
This could lead to a very slow call to __lp_strip_escapes if the gap was large.

The new behavior is to remove escaped sequences from fillchars,
but allow them in the new arguments *fillprefix* and *fillsuffix*.

This actually simplifies the _lp_fill function a bit.

This patch also removes fixed COLUMNS in theme-preview.sh and fix a spacing problem in the unfold theme.