Add extra options to theme-preview.sh for config
The original intent for theme-preview.sh was to have a reference for a
static output of a theme. So all runs of the tool for a specific theme
would always show the same output.

But a different, completely valid use is to see how a theme would look
with your config, or to see how changing your config might change how a
theme looks.

Add optional arguments that will allow for loading the user config,
input files, or to return to the original functionality of locking the
COLUMNS to a set value.

There is still no way to control the mocked data, and there might still
be some weird quirks with the imbedded config in the tool conflicting
with user config.

Fix the minimal theme using the wrong PS1 variable.
Fix Zsh crash in unfold theme.

Fixes #800