Add bug-report.sh
Collect configuration information that is often critical to diagnosing
issues. Most of the issues that need this info end up not being bugs,
but user error or conflicting shell scripts. But at least we should be
able to diagnose those faster.

Add the new script to the bug report template as a required step.