Clarify docs around example config files
The docs assume that tools/config-from-docs.sh will be available. This
may not be the case when installed from a package manager. Add some
clarification that a package manager might have shipped an example
config, where it might be located, and also that the release tarball
will contain an example as well.

Related: #842