Fix DEBUG trap not being cleared on Bash
Bash wants to set the traps of the function instead of the shell level.
You can change this with declare -ft <function>, but I was never able to
get it to work.

Turns out that declare -ft make the function inherit the parent's traps,
not the top level shell. To make this work we just need to climb the
function stack to the top level.

This takes some extra care when running lp_activate() to not error when
clearing PROMPT_COMMAND.

Fixes #849