Add CI testing for each supported shell version
While I frequently do run the test suite against these versions on my
local machine, having them in CI will be faster and will let any
contributor see immediate feedback.

A few tests need tweaking, as they assumed they would not be run by the
root user. While I could change the docker images to not run as root, it
is a bad assumption anyway, so I've removed it.

Curl is needed to install ShUnit2, and ca-certificates is required to
validate the github.com certificate to download it.