Update docs builder versions
Now that Python 3.13 is supported everywhere, might as well update
everything to work on that. Also, why not update all the Sphinx tools at
the same time.

The change in the actual docs is because a warning is issued if you
include the parentheses when calling the `func` class.

The link was excluded as their certificate has been expired for a while.